Crate version_tag

A VersionTag struct for Rust to detect changes

Example

Suppose you want to do a heavy computation based on something else.

You might want to cache the heavy computation result and perform it again only when a dependency changes.

use version_tag::{combine, VersionTag};

/// A dependency
struct Dep {
    v: u32,
    tag: VersionTag,
}

struct MyStruct {
    v: u32,
    tag: VersionTag,
}

impl MyStruct {
    fn update_if_necessary(&mut self, x: &Dep, y: &Dep) {

        // compute the actual tag
        let actual = combine(&[x.tag, y.tag]);

        // if the tag has changed, we need to recalculate
        if actual != self.tag {

            // perform the heavy computation
            self.v = x.v + y.v;

            // update the tag
            self.tag = actual;
        }
    }
}

Structs

VersionTag

Functions

combine
compute a new VersionTag by using the max value of other tags